Separate and cleanup
